requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
appdirs==1.4.4
async-timeout==4.0.2
asyncer==0.0.2
attrs==22.2.0
black==23.1.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
coloredlogs==15.0.1
fastapi==0.87.0
filetype==1.2.0
flake8==6.0.0
Flask==2.2.3
flatbuffers==23.3.3
frozenlist==1.3.3
h11==0.14.0
humanfriendly==10.0
idna==3.4
ImageHash==4.3.1
imageio==2.26.1
importlib-metadata==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
llvmlite==0.39.1
MarkupSafe==2.1.2
mccabe==0.7.0
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.0
numba==0.56.4
numpy==1.23.5
onnxruntime==1.13.1
opencv-python-headless==4.6.0.66
packaging==23.0
pathspec==0.11.1
Pillow==9.3.0
platformdirs==3.2.0
pooch==1.6.0
protobuf==4.22.1
pycodestyle==2.10.0
pydantic==1.10.7
pyflakes==3.0.1
PyMatting==1.1.8
python-multipart==0.0.6
PyWavelets==1.4.1
rembg==2.0.30
requests==2.28.2
requests-toolbelt==0.10.1
scikit-image==0.19.3
scipy==1.9.3
sniffio==1.3.0
starlette==0.21.0
sympy==1.11.1
tifffile==2023.3.21
tomli==2.0.1
tqdm==4.64.1
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.20.0
watchdog==2.1.9
Werkzeug==2.2.3
yarl==1.8.2
zipp==3.15.0
